[Bronze I] Title: 소수가 아닌 수 2, Time: 4 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/29196. 소수가 아닌 수 2/소수가 아닌 수 2.cc
@@ -5,10 +5,10 @@ using namespace std;
 
 int main()
 {
-	float input;
+	double input;
 	cin >> input;
 
-	float p = 1, q = 1;
+	double p = 1, q = 1;
 
 	while (abs((p / q) - input) > 0.000001f) 
 	{
